requirements.txt
python-telegram-bot==20.0
PyPDF2
reportlab
